Refactor tool registration to use modular approach
- Move from inline tool registration to separate registration functions
- Create modular tool files: profiles.ts, avatar-images.ts, experimental.ts
- Add shared API client with utility functions for creating request options
- Rename avatar-utils.ts to avatar-image-api.ts to match OAuth branch naming
- Remove wrapper utility files (profile-utils.ts, experimental-utils.ts)
- Add schemas.ts for centralized tool schema management
- Update index.ts to use new registration functions
- Maintain same 6 MCP tools: get_profile_by_*, get_inferred_interests_by_*, get_avatar_by_*
